+++ cpp/src/CMakeLists.txt
@@ -434,17 +434,17 @@ file(MAKE_DIRECTORY ${root_directory}/releases/include/codegen/third_party/)
 file(MAKE_DIRECTORY ${root_directory}/releases/include/codegen/precompile/)
 file(MAKE_DIRECTORY ${root_directory}/releases/include/codegen/utils/)
 file(MAKE_DIRECTORY ${root_directory}/releases/include/codegen/arrow_compute/ext/)
-file(COPY common/result_iterator.h DESTINATION ${root_directory}/releases/include/common/)
+file(COPY utils/result_iterator.h DESTINATION ${root_directory}/releases/include/common/)
 
 add_definitions(-DNATIVESQL_SRC_PATH="${root_directory}/releases")
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-deprecated-declarations -Wno-attributes")
 set(SPARK_COLUMNAR_PLUGIN_SRCS
         jni/jni_wrapper.cc
         ${PROTO_SRCS}
-        proto/protobuf_utils.cc
-        proto/substrait_utils.cc
-        operators/columnar_to_row_converter.cc
-        shuffle/splitter.cc
+        compute/protobuf_utils.cc
+        compute/substrait_utils.cc
+        operators/c2r/columnar_to_row_converter.cc
+        operators/shuffle/splitter.cc
         )
 
 file(MAKE_DIRECTORY ${root_directory}/releases)
