+++ cpp/src/compute/protobuf_utils.cc
