+++ cpp/src/compute/protobuf_utils.h
