+++ cpp/src/compute/substrait_utils.cc
