+++ cpp/src/utils/exception.h
