+++ cpp/src/utils/result_iterator.h
